# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: mcprack API
  description: >
    Self-service MCP (Model Context Protocol) server catalog and client
    config generator. This API lets an admin manage the server catalog and
    users, and lets any user manage their own server selections, personal
    credential overrides, API tokens, and generated client configs.
  version: "0.0.0"
  contact:
    name: VitexSoftware
    url: https://github.com/VitexSoftware/mcprack
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
  /me:
    get:
      tags: [profile]
      summary: Current user's profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /tokens:
    get:
      tags: [tokens]
      summary: List the current user's API tokens
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Token"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [tokens]
      summary: Create a new API token for the current user
      description: >
        The raw token value is returned exactly once, in this response. It
        cannot be retrieved again — only its hash is persisted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: ci-script
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: "#/components/schemas/Token"
                      - type: object
                        properties:
                          token:
                            type: string
                            description: Raw bearer token — shown only once.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /tokens/{tokenId}:
    delete:
      tags: [tokens]
      summary: Revoke one of the current user's API tokens
      parameters:
        - $ref: "#/components/parameters/TokenId"
      responses:
        "204":
          description: Revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /servers:
    get:
      tags: [servers]
      summary: List servers visible to the current user
      description: Enabled servers the current user is permitted to use.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Server"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/selections:
    get:
      tags: [selections]
      summary: List the current user's selected servers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Selection"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [selections]
      summary: Replace the current user's selected servers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [server_ids]
              properties:
                server_ids:
                  type: array
                  items:
                    type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      server_ids:
                        type: array
                        items:
                          type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/overrides/{serverId}:
    get:
      tags: [overrides]
      summary: Get the current user's override status for a server
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - name: reveal
          in: query
          description: Set to "1" to include decrypted override values.
          schema:
            type: string
            enum: ["1"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Override"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/UpstreamUnavailable"
    put:
      tags: [overrides]
      summary: Set the current user's credential override for a server
      parameters:
        - $ref: "#/components/parameters/ServerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [env]
              properties:
                env:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "200":
          description: OK
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/UpstreamUnavailable"
    delete:
      tags: [overrides]
      summary: Reset a server override back to its default credentials
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "204":
          description: Reset
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/UpstreamUnavailable"
  /me/config/{client}:
    get:
      tags: [config]
      summary: Generate the current user's client config
      parameters:
        - name: client
          in: path
          required: true
          schema:
            type: string
            enum: [claude, copilot]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      filename:
                        type: string
                      config:
                        type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: >
            Unknown client, or nothing to generate (e.g. no servers
            selected, or credentials unavailable).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/servers:
    get:
      tags: [admin-servers]
      summary: List all registered servers (admin)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: include_health
          in: query
          description: Set to "1" to include reachability/credential health per server.
          schema:
            type: string
            enum: ["1"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ServerDetail"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [admin-servers]
      summary: Register a new server (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ServerDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/servers/{serverId}:
    get:
      tags: [admin-servers]
      summary: Get a server's full detail (admin)
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - name: reveal_secrets
          in: query
          description: Set to "1" to include decrypted secret values.
          schema:
            type: string
            enum: ["1"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ServerDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [admin-servers]
      summary: Update a server (admin, partial update)
      parameters:
        - $ref: "#/components/parameters/ServerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ServerDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [admin-servers]
      summary: Delete a server (admin)
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /audit-log:
    get:
      tags: [audit-log]
      summary: List audit trail entries (admin)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: server_id
          in: query
          schema:
            type: integer
        - name: user_id
          in: query
          schema:
            type: integer
        - name: since
          in: query
          description: "YYYY-MM-DDTHH:MM"
          schema:
            type: string
        - name: until
          in: query
          description: "YYYY-MM-DDTHH:MM"
          schema:
            type: string
        - name: errors_only
          in: query
          schema:
            type: string
            enum: ["on"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditLogEntry"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: The session cookie set by POST /login (flask_login).
    bearerAuth:
      type: http
      scheme: bearer
      description: A personal API token minted via POST /api/v1/tokens.
  parameters:
...
    TokenId:
      name: tokenId
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          nullable: true
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        display_name:
          type: string
        auth_type:
          type: string
          enum: [local, ldap]
        is_admin:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
...